Change sounds, add scoreboard
Changed sounds and added a scoreboard that displays how many players on each team are left in game

+++ RoboCat/Inc/RoboCat.h
@@ -60,7 +60,7 @@ private:
 	float				mCatRestitution;
 
 
-	uint32_t			mPlayerId;
+	uint8_t			mPlayerId;
 	std::string			mPlayerTeam;
 
 protected:
